# Various performance optimizations on entity mutations