--- a/src/Entity.js
+++ b/src/Entity.js
@@ -1,11 +1,8 @@
-import { Bitmask } from './utils/Bitmask.js'
-
 export const Entity = (config, registry) => {
   const { entities, systems } = registry
   const removed = []
-  const entityEnabledBitmask = Bitmask(config.maxEntities, Uint32Array)
-
   const deferredEntityRemovals = []
+  const enabledEntities = new Uint8Array(config.maxEntities)
 
   /**
    * Count of entities in this world
@@ -50,7 +47,7 @@ export const Entity = (config, registry) => {
     const eid = removed.length ? removed.shift() : _entityCount
     _entityCount++
     entities.forEach(typedArray => { typedArray[eid] = 0 })
-    entityEnabledBitmask.on(eid)
+    enabledEntities[eid] = 1
     return eid
   }
 
@@ -61,7 +58,7 @@ export const Entity = (config, registry) => {
    */
   const _removeEntity = eid => {
     // Check if entity is already removed
-    if (!entityEnabledBitmask.get(eid)) return
+    if (enabledEntities[eid] === 0) return
 
     // Remove entity from all systems
     for (const s in systems) {
@@ -70,7 +67,7 @@ export const Entity = (config, registry) => {
     }
     removed.push(eid)
     _entityCount--
-    entityEnabledBitmask.off(eid)
+    enabledEntities[eid] = 0
 
     // Clear component bitmasks
     entities.forEach(arr => { arr[eid] = 0 })