--- a/src/Entity.js
+++ b/src/Entity.js
@@ -3,12 +3,7 @@ export const Entity = (config, registry) => {
   const removed = []
   const deferredEntityRemovals = []
   const enabledEntities = new Uint8Array(config.maxEntities)
-
-  /**
-   * Count of entities in this world
-   * @private
-   */
-  let _entityCount = 1
+  let entityCursor = 0
 
   /**
    * Get the count of entities currently in the world.
@@ -23,7 +18,7 @@ export const Entity = (config, registry) => {
    *
    * @memberof module:World
    */
-  const entityCount = () => _entityCount - 1
+  const entityCount = () => entityCursor - removed.length
 
   /**
    * Add a new entity to the world.
@@ -39,38 +34,42 @@ export const Entity = (config, registry) => {
    * @returns {uint32} The entity id.
    */
   const addEntity = () => {
-    if (_entityCount > config.maxEntities) {
+    if (entityCursor >= config.maxEntities) {
       console.warn('❌ Could not add entity, maximum number of entities reached.')
       return
     }
 
-    const eid = removed.length > 0 ? removed.pop() : _entityCount
-    _entityCount++
-    entities.forEach(typedArray => { typedArray[eid] = 0 })
+    const eid = removed.length > 0 ? removed.pop() : entityCursor++
+    for (let arr of entities) arr[eid] = 0
     enabledEntities[eid] = 1
     return eid
   }
 
   /**
    * Internal remove function
-   * @param {uint32} eid
-   * @private true
+   * @private
    */
-  const _removeEntity = eid => {
-    // Check if entity is already removed
-    if (enabledEntities[eid] === 0) return
+  const commitEntityRemovals = () => {
+    if (deferredEntityRemovals.length === 0) return
 
-    // Remove entity from all systems
-    for (const s in systems) {
-      const system = systems[s]
-      system.remove(eid)
+    for (let eid of deferredEntityRemovals) {
+      // Check if entity is already removed
+      if (enabledEntities[eid] === 0) continue
+
+      // Remove entity from all systems
+      for (const s in systems) {
+        systems[s].remove(eid)
+      }
+
+      // Free the entity
+      removed.push(eid)
+      enabledEntities[eid] = 0
+
+      // Clear component bitmasks
+      for (let arr of entities) arr[eid] = 0
     }
-    removed.push(eid)
-    _entityCount--
-    enabledEntities[eid] = 0
 
-    // Clear component bitmasks
-    entities.forEach(arr => { arr[eid] = 0 })
+    deferredEntityRemovals.length = 0
   }
 
   /**
@@ -90,8 +89,8 @@ export const Entity = (config, registry) => {
    * @param {uint32} eid        - The entity id to remove.
    */
   const removeEntity = (eid) => {
-    deferredEntityRemovals.push(() => _removeEntity(eid))
+    deferredEntityRemovals.push(eid)
   }
 
-  return { entityCount, addEntity, removeEntity, deferredEntityRemovals }
+  return { entityCount, addEntity, removeEntity, commitEntityRemovals }
 }