On my benchmark, here are the numbers before:

```
bitecs
  entity_cycle    643 op/s
  add_remove      1,437 op/s
```

And after:

```
bitecs
  entity_cycle    3,323 op/s
  add_remove      6,510 op/s
```

Each commit should okay to cherry-pick if you don't want to merge the whole PR.